Data import api_public_values extraction to presenter and display_name addition
